backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
SQLAlchemy==2.0.34
asyncpg==0.29.0
alembic==1.13.2
redis==5.0.7
httpx==0.27.2
python-dotenv==1.0.1
numpy==2.1.1
pandas==2.2.2
orjson==3.10.7
tenacity==9.0.0
